flask-data/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Marshmallow==1.2.0
marshmallow-sqlalchemy==1.0.0
Flask-Migrate==4.0.7
python-dotenv==1.0.1
flask-cors==4.0.1
